Fix: Issue with clearing the backstack with onboarding. Intent.FLAG_ACTIVITY_CLEAR_TASK was missing due to which we can move back to OnBoardingActivity.kt by back Corrected that. As this exit animation will be used by both of the activity thus moved it to core package.
